- Re-fusionné toutes les modifications faites depuis le branch "port" dans la branche principale; Common.Drawing a été exclu de cette mise à jour.
NE PAS FAIRE UN UPDATE MAINTENANT !



git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2005/Epsitec.Cresus@6519 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/App.DocumentEditor/Dialogs/GlyphArray.cs
@@ -473,7 +473,7 @@ namespace Epsitec.Common.Widgets
 						{
 							if ( this.glyphsMode )
 							{
-								int glyph = (int) this.glyphs[first++];
+								ushort glyph = (ushort) this.glyphs[first++];
 								double width = font.GetGlyphAdvance(glyph)*fontSize;
 								graphics.Rasterizer.AddGlyph(font, glyph, rect.Center.X-width/2.0, rect.Bottom+rect.Height*0.35, fontSize);
 								graphics.RenderSolid(adorner.ColorText(cellState));
@@ -481,7 +481,7 @@ namespace Epsitec.Common.Widgets
 							else
 							{
 								char c = (char) this.unicodes[first++];
-								int glyph = font.GetGlyphIndex(c);
+								ushort glyph = font.GetGlyphIndex(c);
 								double width = font.GetGlyphAdvance(glyph)*fontSize;
 								graphics.AddText(rect.Center.X-width/2.0, rect.Bottom+rect.Height*0.35, c.ToString(), font, fontSize);
 								graphics.RenderSolid(adorner.ColorText(cellState));
